Added some fault checking to the Captain and fixed a bug in the GPS files.

+++ src/GPS_UBX.cpp
@@ -16,7 +16,7 @@ namespace {
 
 namespace GPS_UBX {
 
-long time;
+uint32_t time;
 float longitude;
 float latitude;
 float altitude;
